WIP: introduce an 'io' crate feature
This io feature gates the inclusion of the I/O API provided for
FunctionFS endpoints.

When disabled, consumers can use this crate to describe the ConfigFS
and/or FunctionFS structure of the USB config + functions, and then
bring their own I/O primitives for pushing and pulling data on those
endpoints.

This also yields a dramatic reduction in transitive dependencies pulled
in by this crate, as tokio is no longer always pulled in.

Signed-off-by: Sam Day <me@samcday.com>